+++ backend/src/GraphQl/AppliedMethodInput.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 namespace Database.GraphQl
 {
-    public record AppliedMethodInput(
+    public sealed record AppliedMethodInput(
           Guid MethodId,
           IReadOnlyList<NamedMethodArgumentInput> Arguments,
           IReadOnlyList<NamedMethodSourceInput> Sources