+++ backend/src/GraphQl/FileMetaInformationInput.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Database.GraphQl
 {
-    public record FileMetaInformationInput(
+    public sealed record FileMetaInformationInput(
           string[] Path,
           Guid DataFormatId
     );