+++ backend/src/GraphQl/HygrothermalDataX/CreateHygrothermalDataInput.cs
@@ -5,7 +5,7 @@ using HotChocolate.Types;
 
 namespace Database.GraphQl.HygrothermalDataX
 {
-    public record CreateHygrothermalDataInput(
+    public sealed record CreateHygrothermalDataInput(
           string AccessToken,
           // TODO Why does specifying the type with an attribute not work here?
           [GraphQLType<NonNullType<LocaleType>>] string Locale,