+++ backend/src/Startup.cs
@@ -24,6 +24,8 @@ namespace Database
 {
     public sealed class Startup
     {
+        private const string GraphQlCorsPolicy = "GraphQlCorsPolicy";
+
         private readonly IWebHostEnvironment _environment;
         private readonly AppSettings _appSettings;
 
@@ -44,6 +46,10 @@ namespace Database
             ConfigureMessageSenderServices(services);
             ConfigureRequestResponseServices(services);
             ConfigureSessionServices(services);
+            services.AddAntiforgery(_ =>
+            {
+                _.HeaderName = "X-XSRF-TOKEN";
+            });
             services
                 .AddDataProtection()
                 .PersistKeysToDbContext<Data.ApplicationDbContext>();
@@ -70,11 +76,13 @@ namespace Database
             }
             );
             services.AddCors(_ =>
-                _.AddDefaultPolicy(policy =>
-                    policy
-                    .AllowAnyOrigin()
-                    .AllowAnyHeader()
-                    .AllowAnyMethod()
+                _.AddPolicy(
+                     name: GraphQlCorsPolicy,
+                     policy =>
+                        policy
+                        .AllowAnyOrigin()
+                        .AllowAnyHeader()
+                        .AllowAnyMethod()
                     )
                 );
             services.AddControllersWithViews();
@@ -170,7 +178,7 @@ namespace Database
             // app.UseHttpsRedirection(); // Done by NGINX
             app.UseSerilogRequestLogging();
             app.UseStaticFiles();
-            app.UseCookiePolicy();
+            app.UseCookiePolicy();  // [SameSite cookies](https://learn.microsoft.com/en-us/aspnet/core/security/samesite)
             app.UseRouting();
             // TODO Do we really want this? See https://docs.microsoft.com/en-us/aspnet/core/fundamentals/localization?view=aspnetcore-5.0
             app.UseRequestLocalization(_ =>
@@ -189,7 +197,8 @@ namespace Database
             /* app.UseWebSockets(); */
             app.UseEndpoints(_ =>
             {
-                _.MapGraphQL().WithOptions(
+                _.MapGraphQL()
+                .WithOptions(
                     // https://chillicream.com/docs/hotchocolate/server/middleware
                     new GraphQLServerOptions
                     {
@@ -206,7 +215,8 @@ namespace Database
                             Title = "GraphQL"
                         }
                     }
-                );
+                )
+                .RequireCors(GraphQlCorsPolicy);
                 _.MapControllers();
                 _.MapHealthChecks("/health",
                     new HealthCheckOptions